audio_deepfake_eval/requirements.txt
# Core dependencies
numpy>=1.20.0
scikit-learn>=0.24.0
PyYAML>=5.4.1

# Visualization dependencies
matplotlib>=3.3.0
seaborn>=0.11.0 